DE_ICE: rewrite to use an intermediary map; handle .trip.stops being absent
If iceportal.de is having a weird day, .trip.stops in tripInfo may be null,
which would cause that entire block to throw an error.

I originally did this with just string concatenation, but an intermediary
map like this feels less dirty and makes the code a bit more readable.